+++ ui/src/main/java/com/jin/ui/review/write/content/MenuReviewWriteScreen.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.reviewwrite.ui.content
+package com.jin.ui.review.write.content
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
@@ -31,7 +31,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.domain.cart.model.Cart
 import com.jin.domain.review.ReviewContent
 import com.jin.ui.theme.PointColor
